feat: mLSTM Triton Kernels (#79)
* feat: support mutable variables

* feat: option for logging logit stats

* chore: add tests for intermediates

* feat: adding intermediates to xlstm

* fix: some test fixes and new backend compare

* fix: correcting weird precommit op

* chore: some more docs

* fix: more docs

* feat: adding fwbw backend, tests currently failing

* fix: working fwbw jax version, but diff to parallel in pytorch as well

* fix: working pytorch-jax match

* chore: add tests for stabilized and unstabilized

* fix: add backward test and corrections

* feat: removing head axis in jax fwbw

* feat: adding scan in fwbw forward, might not be optimal yet

* chore: adding pytorch dtypes in simple

* feat: add mlstm as submodule

* fix: add submodule

* feat: fw kernel running, but tests fail

* feat: add triton bw, to be tested

* feat: implementing bw and fwbw

* chore: add passing fwbw tests against PyTorch

* feat: integrating kernels in model, still compilation issues

* feat: running training with mixed precision

* feat: running eval

* chore: doc strings for fwbw

* chore: add doc strings everywhere

* chore: remove dtype in doc strings

* chore: use more efficient stride

* chore: clean up slimpajama script

* chore: doc strings for tests

* chore: more doc strings

* wip: add pytorch kernel script and debug data pipeline

* fix: addressable shards

* fix: correct kernel bug and add comparison script

* feat: adding 165M model config

* chore: add commenting on kernel scripts

* chore: fix tests and more comments

* chore: some more docs in backends

* chore: improved docs

* chore: add back tests

* fix: reduce backend tests for faster CI actions

* feat: kernel optimizations

* chore: addressing comments

* chore: refactor tests into fixture

* fix: backend tests

* chore: remove kernel optimization for later PR

+++ xlstm_jax/models/xlstm_pytorch/blocks/mlstm/backend/simple.py
@@ -50,7 +50,7 @@ def parallel_stabilized_simple(
 
     log_fgates_cumsum = torch.cat(
         [
-            torch.zeros((B, NH, 1, 1), dtype=_dtype, device=_device),
+            torch.zeros((B, NH, 1, 1), dtype=log_fgates.dtype, device=_device),
             torch.cumsum(log_fgates, dim=-2),
         ],
         dim=-2,
@@ -76,13 +76,16 @@ def parallel_stabilized_simple(
         # (B, NH, 1, 1)
     log_D_matrix_stabilized = log_D_matrix - max_log_D  # (B, NH, S, S)
     D_matrix = torch.exp(log_D_matrix_stabilized)  # (B, NH, S, S)
+    D_matrix = D_matrix.to(_dtype)
 
     keys_scaled = keys / math.sqrt(DH)
 
     # combination matrix C
     qk_matrix = queries @ keys_scaled.transpose(-2, -1)  # (B, NH, S, S)
     C_matrix = qk_matrix * D_matrix  # (B, NH, S, S)
-    normalizer = torch.maximum(C_matrix.sum(dim=-1, keepdim=True).abs(), torch.exp(-max_log_D))  # (B, NH, S, 1)
+    normalizer = torch.maximum(
+        C_matrix.sum(dim=-1, keepdim=True).abs(), torch.exp(-max_log_D).to(_dtype)
+    )  # (B, NH, S, 1)
     # (B, NH, S, S)
     C_matrix_normalized = C_matrix / (normalizer + eps)
